PangoWin32: Query font description using DirectWrite
We now use DirectWrite to query the font weight, stretch, description
and so on by using first DirectWrite's GdiInterop to convert the
logfontw into an IDirectWriteFont, since the support in there for
querying font attributes are more complete in there.

Portions based on Luca Bacci's work for querying the font description
for Windows using DirectWrite for the upcoming Pango2.